Switches to Google Login instead of Google+
Google+ is being deprecated, need to switch.

Following https://developers.google.com/identity/sign-in/web/sign-in.

The frontend change is straight forward taken from the Google
example. For it to work in Firefox I had to clear my cache, though.

In the backend the Google Library is used to load the email from a
token and then pass it forward to Spring security. This allows us to
remove our custom JSON mapping classes!