Wraps WebSecurityConfigurerAdapter configuration classes
The JAR file from the previous commit did not start because of
circular dependencies (?).

To fix this the UserDetailsManager Bean declaration had to be pulled
to another level.

So now there is a SecurityConfiguration class that wraps the two
WebSecurityConfigurerAdapter implementations for google-login and form
login as inner classes while containing the UserDetailsManager on a
higher level.